backend/repository/cache/redis/login_test.go
package redis

import (
	"context"
	"testing"

	"chat-app-demo/domain"
	"chat-app-demo/lib"
	"chat-app-demo/repository/db"

	"github.com/redis/go-redis/v9"
)

func TestLogin(t *testing.T) {
	defer DeleteAll()

	rc := NewRedisCache()

	// Prepare data
	dbrepo, err := db.NewDBRepository()
	if err != nil {
		t.Fatal(err)
	}
	defer dbrepo.Close()

	uObj, _ := domain.NewUser("u1")
	err = dbrepo.AddUser(*uObj)
	if err != nil {
		t.Fatal(err)
	}

	lObj := domain.NewLogin("u1")
	err = dbrepo.AddLogin(*lObj)
	if err != nil {
		t.Fatal(err)
	}

	// Get data
	flObj, err := rc.FindLogin("u1")
	if err != nil {
		t.Fatal(err)
	}
	if !lib.CompareStructs(*lObj, *flObj) {
		t.Fatal("doesn't match")
	}

	// Confirm it's cached
	_, err = rc.rdb.Get(context.Background(), keyLogin("u1")).Result()
	if err != nil {
		t.Fatal(err)
	}

	// Delete cache
	err = rc.DeleteLogin("u1")
	if err != nil {
		t.Fatal(err)
	}

	// Confirm cached data is deleted
	_, err = rc.rdb.Get(context.Background(), keyLogin("u1")).Result()
	if err != redis.Nil {
		t.Fatal("Expected redis.Nil, got:", err)
	}
}
